Clean up all analysers.
