+++ certstream_analytics/analysers/__init__.py
@@ -1,7 +1,8 @@
 # pylint: disable=missing-docstring
 from .base import Analyser, Debugger
-from .common_domain_analyser import AhoCorasickDomainMatching, WordSegmentation
-from .common_domain_analyser import DomainMatchingOption, DomainMatching
+from .domain_matching import AhoCorasickDomainMatching
+from .domain_matching import DomainMatchingOption, DomainMatching
+from .common_domain_analyser import WordSegmentation
 from .common_domain_analyser import BulkDomainMarker
 from .common_domain_analyser import FeaturesGenerator
 from .common_domain_analyser import IDNADecoder
