+++ certstream_analytics/analysers/common_domain_analyser.py
@@ -1,115 +1,22 @@
 '''
-Verify the domain against the list of most popular domains from OpenDNS
-(https://github.com/opendns/public-domain-lists). Let's see how useful
-it is to prevent phishing domains.
+The list of basic analysers includes:
+    - WordSegmentation
+    - IDNADecoder
+    - HomoglyphsDecoder
+    - FeaturesGenerator (generate various features for further downstream processing)
+    - BulkDomainMarker
 '''
-from enum import Enum
-
 import re
 import tldextract
 import wordsegment
 from nostril import nonsense
 import idna
 from confusable_homoglyphs import confusables
-import ahocorasick
 
 from .base import Analyser
 
-# Take a histogram here and find out the suitable value for this
-BULK_DOMAIN_THRESHOLD = 15
-
 
 # pylint: disable=too-few-public-methods
-class AhoCorasickDomainMatching(Analyser):
-    '''
-    The domain and its SAN will be compared against the list of domains, for
-    example, the most popular domains from OpenDNS.
-    '''
-    # Get this number from the histogram of the length of all top domains
-    MIN_MATCHING_LENGTH = 3
-
-    # Some domains that don't work too well with tldextract and generate too
-    # many FPs
-    EXCLUDED_DOMAINS = {
-        'www': 1,
-        'web': 1,
-    }
-
-    # Some common domain parts that cause too many FP
-    IGNORED_PARTS = r'^(autodiscover\.|cpanel\.)'
-
-    def __init__(self, domains):
-        '''
-        Use Aho-Corasick to find the matching domain so we construct its Trie
-        here. Thought: How the f**k is com.com in the list?
-        '''
-        self.automaton = ahocorasick.Automaton()
-        self.domains = {}
-
-        for index, domain in enumerate(domains):
-            # Processing only the domain part.  All sub-domains or TLDs will
-            # be ignored, for example:
-            #   - www.google.com becomes google
-            #   - www.google.co.uk becomes google
-            #   - del.icio.us becomes icio
-            ext = tldextract.extract(domain)
-
-            if ext.domain in AhoCorasickDomainMatching.EXCLUDED_DOMAINS:
-                continue
-
-            self.automaton.add_word(ext.domain, (index, ext.domain))
-            self.domains[ext.domain] = domain
-
-        self.automaton.make_automaton()
-
-    def run(self, record):
-        '''
-        Use Aho-Corasick to find the matching domain. Check the time complexity
-        of this function later.
-
-        Tricky situation #1: When the string (domain) in the Trie is too short,
-        it could match many domains, for example, g.co or t.co.  So they need
-        to be ignored somehow.  Looking at the histogram of the length of all
-        domains in the list, there are only less than 100 domains with the
-        length of 2 or less.  So we choose to ignore those.  Also, we will
-        prefer longer match than a shorter one for now.
-        '''
-        if 'analysers' not in record:
-            record['analysers'] = []
-
-        results = {}
-        # Check the domain and all its SAN
-        for domain in record['all_domains']:
-            # Remove wildcard
-            domain = re.sub(r'^\*\.', '', domain)
-
-            # Remove some FP-prone parts
-            domain = re.sub(AhoCorasickDomainMatching.IGNORED_PARTS, '', domain)
-
-            # Similar to all domains in the list, the TLD will be stripped off
-            ext = tldextract.extract(domain)
-            # The match will be a tuple in the following format: (5, (0, 'google'))
-            matches = [m[1][1] for m in self.automaton.iter('.'.join(ext[:2]))
-                       if len(m[1][1]) >= AhoCorasickDomainMatching.MIN_MATCHING_LENGTH]
-
-            if matches:
-                matches.sort(key=len)
-
-                match = matches[-1]
-                # We only keep the the longest match of the first matching domain
-                # for now
-                results[domain] = [self.domains[match]] if match in self.domains else match
-                break
-
-        if results:
-            record['analysers'].append({
-                'analyser': type(self).__name__,
-                'output': results,
-            })
-
-        return record
-
-
 class WordSegmentation(Analyser):
     '''
     Perform word segmentation of all the SAN domains as an attempt to make sense
@@ -159,7 +66,7 @@ class WordSegmentation(Analyser):
         results = {}
         # Check the domain and all its SAN
         for domain in record['all_domains']:
-            # Remove wildcard
+            # Remove wild card
             domain = re.sub(r'^\*\.', '', domain)
 
             # The TLD will be stripped off cause it does not contribute anything here
@@ -191,134 +98,6 @@ class WordSegmentation(Analyser):
         return record
 
 
-class DomainMatchingOption(Enum):
-    '''
-    Control how strict we want to do our matching.
-    '''
-    # For example applefake.it will match with apple.com case ['apple'] is
-    # a subset of ['apple', 'fake']
-    SUBSET_MATCH = 0
-
-    # Similar but use in instead of issubset so that the order is preserved
-    ORDER_MATCH = 1
-
-
-class DomainMatching(Analyser):
-    '''
-    This is the first example of the new group of meta analysers which are used
-    to combine the result of other analysers.
-    '''
-    def __init__(self, include_tld=True, option=DomainMatchingOption.ORDER_MATCH):
-        '''
-        Just load the wordsegment package, whatever it is.
-        '''
-        wordsegment.load()
-
-        # Save the matching option here so we can refer to it later
-        self.include_tld = include_tld
-
-        self.option = {
-            DomainMatchingOption.SUBSET_MATCH: set,
-            DomainMatchingOption.ORDER_MATCH: list,
-        }[option]
-
-    def run(self, record):
-        '''
-        Note that a meta-analyser will need to run after other analysers have
-        finished so that their outputs are available.
-        '''
-        if 'analysers' not in record:
-            return record
-
-        analysers = {
-            AhoCorasickDomainMatching.__name__: {},
-            WordSegmentation.__name__: {},
-            BulkDomainMarker.__name__: {},
-        }
-
-        for analyser in record['analysers']:
-            name = analyser['analyser']
-
-            if name not in analysers:
-                continue
-
-            if name == BulkDomainMarker.__name__ and analyser['output']:
-                # Skip bulk record and deal with it later, with such large
-                # number of SAN name, it's bound to be a match
-                continue
-
-            analysers[name] = analyser['output']
-
-        # Check that all outputs are there before continuing
-        if not analysers[AhoCorasickDomainMatching.__name__] or not analysers[WordSegmentation.__name__]:
-            return record
-
-        results = self._match(analysers[AhoCorasickDomainMatching.__name__],
-                              analysers[WordSegmentation.__name__])
-
-        if results:
-            record['analysers'].append({
-                'analyser': type(self).__name__,
-                'output': results,
-            })
-
-        return record
-
-    def _match(self, ahocorasick_output, segmentation_output):
-        '''
-        Use internally by the run function to combine AhoCorasick and WordSegmentation
-        results.
-        '''
-        results = {}
-        # Check all the matching domains reported by AhoCorasick analyser
-        for match, domains in ahocorasick_output.items():
-            # The result of AhoCorasick matcher is a list of matching domains, for example,
-            #
-            #   {
-            #       'analyser': 'AhoCorasickDomainMatching',
-            #       'output': {
-            #           'login-appleid.apple.com.managesuppport.co': ['apple.com', 'support.com'],
-            #       },
-            #   },
-            #
-            if match not in segmentation_output:
-                continue
-
-            phish = self.option(segmentation_output[match])
-            match_ext = tldextract.extract(match)
-
-            for domain in domains:
-                ext = tldextract.extract(domain)
-
-                # This record is from a legitimate source, for example, agrosupport.zendesk.com
-                # will match with zendesk.com. In our case, we don't really care about this so
-                # it will be ignored and not reported as a match.
-                if ext[1:] == match_ext[1:]:
-                    continue
-
-                tmp = []
-                # Intuitively, it will be more accurate if we choose to include the TLD here.
-                # For example, if both 'apple' and 'com' appear in the matching domain, it's
-                # very likely that something phishing is going on here. On the other hand,
-                # if only 'apple' occurs, we are not so sure and it's better left for more
-                # advance analysers to have their says in that
-                for part in ext[:] if self.include_tld else ext[:2]:
-                    for token in part.split('.'):
-                        tmp.extend(wordsegment.segment(token))
-
-                legit = self.option(tmp)
-
-                if (isinstance(phish, set) and legit.issubset(phish)) or \
-                   (isinstance(phish, list) and '.{}'.format('.'.join(legit)) in '.'.join(phish)):
-                    # Found a possible phishing domain
-                    if match not in results:
-                        results[match] = []
-
-                    results[match].append(domain)
-
-        return results
-
-
 class BulkDomainMarker(Analyser):
     '''
     Mark the record that has tons of SAN domains in it. Most of the time, they are
@@ -327,7 +106,10 @@ class BulkDomainMarker(Analyser):
     verify the similarity among these domains. A lower similarity score means these
     domains are totally unrelated.
     '''
-    def __init__(self, threshold=BULK_DOMAIN_THRESHOLD):
+    # Take a histogram here and find out the suitable value for this
+    THRESHOLD = 15
+
+    def __init__(self, threshold=THRESHOLD):
         '''
         Set the threshold to mark the record as a bulk record.
         '''
@@ -342,7 +124,7 @@ class BulkDomainMarker(Analyser):
         if 'analysers' not in record:
             record['analysers'] = []
 
-        is_bulked = True if len(record['all_domains']) >= self.threshold else False
+        is_bulked = len(record['all_domains']) >= self.threshold
 
         record['analysers'].append({
             'analyser': type(self).__name__,
@@ -406,6 +188,24 @@ class HomoglyphsDecoder(Analyser):
         '''
         self.greedy = greedy
 
+    @staticmethod
+    def is_latin(alt):
+        '''
+        Check if a string is in Latin cause, in our specific case, we will
+        only care about Latin characters
+        '''
+        lower_s = range(ord('a'), ord('z') + 1)
+        upper_s = range(ord('A'), ord('Z') + 1)
+
+        # We need to check the length of the homoglyph here cause
+        # confusable_homoglyphs library nicely returns multi-character
+        # match as well, for example, 'rn' has an alternative of 'm'
+        for alt_c in alt:
+            if ord(alt_c) not in lower_s and ord(alt_c) not in upper_s:
+                return False
+
+        return True
+
     def run(self, record):
         '''
         Using the confusable-homoglyphs, we are going to generate all alternatives ASCII
@@ -413,16 +213,12 @@ class HomoglyphsDecoder(Analyser):
         '''
         decoded = []
 
-        # For our specific case, we will only care about latin character
-        lower_s = range(ord('a'), ord('z') + 1)
-        upper_s = range(ord('A'), ord('Z') + 1)
-
         for domain in record['all_domains']:
             wildcard = False
 
             if re.match(r'^\*\.', domain):
                 wildcard = True
-                # Remove wildcard to simplify the domain name a bit and we'll put it back later
+                # Remove wild card to simplify the domain name a bit and we'll put it back later
                 domain = re.sub(r'^\*\.', '', domain)
 
             hg_map = {hg['character']: hg for hg in confusables.is_confusable(domain, greedy=True)}
@@ -430,7 +226,7 @@ class HomoglyphsDecoder(Analyser):
 
             for domain_c in domain:
                 # Confusable homoglyphs could not find any homoglyphs for this character
-                # so we decice to keep the original character as it is
+                # so we decide to keep the original character as it is
                 if domain_c not in hg_map:
                     decoded_domain_c.append([domain_c])
                     continue
@@ -439,20 +235,11 @@ class HomoglyphsDecoder(Analyser):
                 hglyph = hg_map[domain_c]
 
                 if hglyph['alias'] == 'LATIN':
-                    # The character is latin, we don't need to do anything here
+                    # The character is Latin, we don't need to do anything here
                     found.append(hglyph['character'])
 
                 for alt in hglyph['homoglyphs']:
-                    is_latin = True
-                    # We need to check the lengh of the homoglyph here cause
-                    # confusable_homoglyphs library nicely returns multi-character
-                    # match as well, for example, 'rn' has an alternative of 'm'
-                    for alt_c in alt['c']:
-                        if ord(alt_c) not in lower_s and ord(alt_c) not in upper_s:
-                            is_latin = False
-                            break
-
-                    if is_latin:
+                    if HomoglyphsDecoder.is_latin(alt['c']):
                         found.append(alt['c'].lower())
 
                 # If nothing is found, we keep the original character
@@ -553,7 +340,7 @@ class FeaturesGenerator(Analyser):
                 x.append(randomness_count / len(segments))
 
                 x_samples.append(x)
-                Y_samples.append(True if 'usual_suspect' in record else False)
+                Y_samples.append('usual_suspect' in record)
 
             break
