+++ setup.py
@@ -17,8 +17,24 @@ setup(
     license='MIT',
     long_description=long_description,
     long_description_content_type='text/markdown',
-    install_requires=['elasticsearch_dsl', 'certstream', 'pyahocorasick', 'tldextract', 'wordsegment', 'pyenchant', 'idna'],
-    tests_require=['unittest', 'coverage', 'nose', 'pytest-pep8', 'pytest-cov', 'codecov'],
+    install_requires=[
+        'elasticsearch_dsl',
+        'certstream',
+        'pyahocorasick',
+        'tldextract',
+        'wordsegment',
+        'pyenchant',
+        'idna', 
+        'confusable_homoglyphs'
+    ],
+    tests_require=[
+        'unittest',
+        'coverage',
+        'nose',
+        'pytest-pep8',
+        'pytest-cov',
+        'codecov'
+    ],
     dependency_links=[
         'https://github.com/casics/nostril/tarball/master'
     ],
